+++ rr/openpgpkey.js
@@ -30,7 +30,8 @@ export default class OPENPGPKEY extends RR {
   /******  IMPORTERS   *******/
   fromBind({ bindline: bindline }) {
     // test.example.com  3600  IN  OPENPGPKEY  <base64 public key>
-    const regex = /^(?<owner>\S+)\s+(?<ttl>\d{1,10})\s+(?<class>IN)\s+(?<type>OPENPGPKEY)\s+(?<publickey>\S[\s\S]*)$/i;
+    const regex =
+      /^(?<owner>\S+)\s+(?<ttl>\d{1,10})\s+(?<class>IN)\s+(?<type>OPENPGPKEY)\s+(?<publickey>\S[\s\S]*)$/i
     const match = bindline.trim().match(regex)
     if (!match) this.throwHelp(`unable to parse OPENPGPKEY: ${bindline}`)