+++ test/cert.js
@@ -3,6 +3,8 @@ import * as base from './base.js'
 
 import CERT from '../rr/cert.js'
 
+const defaults = { class: 'IN', ttl: 3600, type: 'CERT' }
+
 const validRecords = [
   {
     owner: 'mail.example.com.',
@@ -30,7 +32,44 @@ const validRecords = [
   },
 ]
 
-const invalidRecords = []
+const invalidRecords = [
+  {
+    ...defaults,
+    owner: 'mail.example.com.',
+    'cert type': 'INVALID', // not a valid mnemonic
+    'key tag': 0,
+    algorithm: 0,
+    certificate: 'hexidecimalkeystring1',
+    msg: /unknown cert type mnemonic/i,
+  },
+  {
+    ...defaults,
+    owner: 'mail.example.com.',
+    'cert type': 'PGP',
+    'key tag': 70000, // out of 16-bit range
+    algorithm: 0,
+    certificate: 'hexidecimalkeystring1',
+    msg: /16-bit integer/i,
+  },
+  {
+    ...defaults,
+    owner: 'mail.example.com.',
+    'cert type': 'PGP',
+    'key tag': 0,
+    algorithm: 300, // out of 8-bit range
+    certificate: 'hexidecimalkeystring1',
+    msg: /8-bit integer/i,
+  },
+  {
+    ...defaults,
+    owner: 'mail.example.com.',
+    'cert type': 'PGP',
+    'key tag': 0,
+    algorithm: 0,
+    certificate: '', // empty certificate
+    msg: /certificate is required/i,
+  },
+]
 
 describe('CERT record', function () {
   base.valid(CERT, validRecords)